Version 1.2.14
03.05.2002 - Added option execd, which works just like exec only
             that it removes the original file after executing the
             command.
05.05.2002 - Dialog dir_ctrl now shows the number of files and bytes
             in the directories and queue that are to be monitored.
             the structure) changes.
13.05.2002 - Specifying a group name together with a hostname was not
             handled correctly. Fixed.
14.05.2002 - Reduce priority when password is wrong.
16.05.2002 - Removed all shared memory stuff from AFD.
             command failes. This only works without the %s parameter.
22.05.2002 - Separate old file times for unknown and queued files in
             dir_options.
27.05.2002 - Combine FTP rename to one command, not two separate ones.
10.06.2002 - Not specifying a filename for the assemble now gives a
             warning in the receive log.
11.06.2002 - The field dir_alias was not set by FD, after a DIR_CONFIG
             update this could lead to data being written to the end
             of the FRA. Fixed.
14.06.2002 - Assemble MSS did not work. Fixed.
             an error.
16.06.2002 - Added option unix time to rename rule.
29.06.2002 - New options toupper and tolower to convert all letters of
             the file name to upper or lower case.
02.07.2002 - Added from option for mail.
03.07.2002 - When reordering the hostnames in afd_ctrl dialog with
             edit_hc it could happen, due to a bug in function
             create_db(), that the files of some directories are just
             deleted. Since there was no reliable way to reproduce this,
             if this fix corrected it. Only time will tell.
04.07.2002 - For mail add To header by default.
             correctly.
05.07.2002 - FD sometimes crashes when some hosts are being removed
             from the FSA. Fixed.
08.07.2002 - Removed combination of FTP rename to one command, it caused
             to many problems with window systems.

+++ X/Motif/show_queue/delete_files.c
@@ -39,6 +39,7 @@ DESCR__S_M3
  **
  ** HISTORY
  **   29.07.2001 H.Kiehl Created
+ **   06.05.2002 H.Kiehl Remove files and bytes from FRA queue data.
  **
  */
 DESCR__E_M3
@@ -76,13 +77,18 @@ extern struct delete_log       dl;
 #endif
 
 /* Global variables */
-int                            fsa_fd = -1,
+int                            fra_fd = -1,
+                               fra_id,
+                               fsa_fd = -1,
                                fsa_id,
+                               no_of_dirs = 0,
                                no_of_hosts,
                                counter_fd;
 #ifndef _NO_MMAP
-off_t                          fsa_size;
+off_t                          fra_size,
+                               fsa_size;
 #endif
+struct fileretrieve_status     *fra;
 struct filetransfer_status     *fsa;
 struct afd_status              *p_afd_status;
 
@@ -217,6 +223,16 @@ delete_files(int no_selected, int *select_list)
       }
    }
 
+   if (toggles_set & SHOW_INPUT)
+   {
+      /* Map to the FRA. */
+      if (fra_attach() == INCORRECT)
+      {
+         (void)fprintf(stderr, "Failed to attach to FRA.\n");
+         exit(INCORRECT);
+      }
+   }
+
    for (i = 0; i < no_selected; i++)
    {
       if ((qfl[select_list[i] - 1].queue_type == SHOW_OUTPUT) &&
@@ -292,7 +308,7 @@ delete_files(int no_selected, int *select_list)
                *ptr = '\0';
                if (rmdir(fullname) == -1)
                {
-                  if (errno != ENOTEMPTY)
+                  if ((errno != ENOTEMPTY) && (errno != EEXIST))
                   {
                      (void)xrec(toplevel_w, INFO_DIALOG,
                                 "Failed to rmdir() <%s> : %s (%s %d)",
@@ -378,13 +394,26 @@ delete_files(int no_selected, int *select_list)
                                qfl[select_list[i] - 1].hostname,
                                qfl[select_list[i] - 1].file_name);
               }
-              if (unlink(fullname) != -1)
+              if (unlink(fullname) == -1)
               {
-                 deleted = YES;
+                 deleted = NO;
               }
               else
               {
-                 deleted = NO;
+                 if (qfl[select_list[i] - 1].hostname[0] != '\0')
+                 {
+                    int k;
+
+                    for (k = 0; k < no_of_dirs; k++)
+                    {
+                       if (qfl[select_list[i] - 1].dir_id_pos == fra[k].dir_pos)
+                       {
+                          ABS_REDUCE_QUEUE(k, 1, qfl[select_list[i] - 1].size);
+                          break;
+                       }
+                    }
+                 }
+                 deleted = YES;
               }
            }
 
@@ -478,6 +507,10 @@ delete_files(int no_selected, int *select_list)
       }
       (void)fsa_detach(NO);
       (void)detach_afd_status();
+   }
+   if (toggles_set & SHOW_OUTPUT)
+   {
+      (void)fra_detach();
    }
 
    /* Tell user what we have done. */
